Adding Git Style Messages
Like git when a command that can take a note as an agrument does note
include the note the user is prompted with their default editior to add
or edit the existing note depending on the command used.

If start is used the note is defaulted to blank if left blank no note is
set.

If stop the existing note is used to populate the editor.  Changes to
this note will overwrite the existing note, not append